fix: Handle reset for old browser
We rely on the `indexedDb.databases()` method to iterate over all local
db to be removed, at login time. We also do it later, in the
`SharedWorkerProvider` as a fallback.
However, this does not work for old browsers that do not support this
function.
Therefore, we now store the db names locally in PouchLink, and rely on
it to remove the databases.

+++ src/dataproxy/worker/data.ts
@@ -27,23 +27,29 @@ export const queryIsTrustedDevice = async (
 }
 
 const deleteDatabases = async (): Promise<void> => {
-  const databases = await window.indexedDB.databases()
-  // Remove all indexedDB databases
-  for (const db of databases) {
-    if (db.name) {
-      window.indexedDB.deleteDatabase(db.name)
-      log.info('Deleted indexedDB database : ', db.name)
+  if (typeof window.indexedDB?.databases === 'function') {
+    const databases = await window.indexedDB.databases()
+    // Remove all indexedDB databases
+    for (const db of databases) {
+      if (db.name) {
+        window.indexedDB.deleteDatabase(db.name)
+        log.info('Deleted indexedDB database : ', db.name)
+      }
     }
   }
 }
 
 export const removeStaleLocalData = async (): Promise<void> => {
   // Check flag existence proving the reset process was incomplete
-  const hasStaleData = localStorage.getItem(LOCALSTORAGE_KEY_DELETING_DATA)
-  if (hasStaleData) {
-    log.info('Found stale data: remove it')
-    await deleteDatabases()
-    localStorage.removeItem(LOCALSTORAGE_KEY_DELETING_DATA)
+  try {
+    const hasStaleData = localStorage.getItem(LOCALSTORAGE_KEY_DELETING_DATA)
+    if (hasStaleData) {
+      log.info('Found stale data: remove it')
+      await deleteDatabases()
+      localStorage.removeItem(LOCALSTORAGE_KEY_DELETING_DATA)
+    }
+  } catch (e) {
+    log.error(e)
   }
   return
 }
